+++ certbot/certbot/_internal/renewal.py
@@ -47,7 +47,7 @@ STR_CONFIG_ITEMS = ["config_dir", "logs_dir", "work_dir", "user_agent",
                     "server", "account", "authenticator", "installer",
                     "renew_hook", "pre_hook", "post_hook", "http01_address",
                     "preferred_chain", "key_type", "elliptic_curve"]
-INT_CONFIG_ITEMS = ["rsa_key_size", "http01_port"]
+INT_CONFIG_ITEMS = ["rsa_key_size", "http01_port", "certificate_validity"]
 BOOL_CONFIG_ITEMS = ["must_staple", "allow_subset_of_names", "reuse_key",
                      "autorenew"]